backend/requirements.txt
# Creator Analytics SaaS - Backend Requirements

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Redis & Caching
redis>=5.0.1
celery>=5.3.6

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Validation & Settings
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client
httpx>=0.26.0
aiofiles>=23.2.1

# Utilities
python-dateutil>=2.8.2
PyYAML>=6.0.1
tenacity>=8.2.3
structlog>=24.1.0

# Data Processing
numpy>=1.26.3
pandas>=2.1.4

# Machine Learning (CPU-only for Render)
torch>=2.1.2
transformers>=4.36.2

# AI & NLP
openai>=1.12.0

# Monitoring
prometheus-client>=0.19.0

# AWS S3
boto3>=1.34.1

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0
